pyproject.toml
[project]
name = "multimodal-ecommerce-qa"
version = "0.1.0"
description = "Text-to-image retrieval baseline for the ecommerce multimodal RAG assignment."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
  "accelerate>=1.10.1",
  "bitsandbytes>=0.48.1",
  "numpy>=2.2.4",
  "openai>=2.0.0",
  "pillow>=11.2.1",
  "pypdf>=5.4.0",
  "qwen-vl-utils>=0.0.11",
  "torch==2.8.0",
  "torchvision==0.23.0",
  "tqdm>=4.67.1",
  "transformers[serving]==5.5.0",
]

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ecom_rag"]

[tool.uv]
package = true

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git", rev = "main" }
